refactor: package 구조 변경
