+++ app/src/main/java/com/teamwss/websoso/ui/notice/adapter/NoticeAdapter.kt
@@ -1,4 +1,4 @@
-package com.teamwss.websoso.ui.notice
+package com.teamwss.websoso.ui.notice.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
