+++ app/src/main/res/layout/item_notice.xml
@@ -22,7 +22,7 @@
             app:layout_constraintDimensionRatio="1"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            tools:src="@mipmap/ic_launcher" />
+            tools:src="@drawable/ic_home_alert" />
 
         <TextView
             android:id="@+id/tv_item_notice_title"
